+++ src/main/java/kitchenpos/menus/application/dto/ProductDto.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.application;
+package kitchenpos.menus.application.dto;
 
 import java.util.UUID;
