+++ src/main/java/kitchenpos/menus/infra/DefaultProductApiService.java
@@ -7,7 +7,7 @@ import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 
 import kitchenpos.menus.application.ProductApiService;
-import kitchenpos.menus.application.ProductDto;
+import kitchenpos.menus.application.dto.ProductDto;
 import kitchenpos.menus.domain.Price;
 import kitchenpos.products.domain.Product;
 import kitchenpos.products.domain.ProductRepository;
