+++ src/main/java/kitchenpos/menus/ui/MenuRestController.java
@@ -1,7 +1,7 @@
 package kitchenpos.menus.ui;
 
-import kitchenpos.menus.application.ChangeMenuPriceRequest;
-import kitchenpos.menus.application.CreateMenuRequest;
+import kitchenpos.menus.application.dto.ChangeMenuPriceRequest;
+import kitchenpos.menus.application.dto.CreateMenuRequest;
 import kitchenpos.menus.application.MenuService;
 import kitchenpos.menus.domain.Menu;
 import org.springframework.http.ResponseEntity;
