+++ src/main/java/kitchenpos/products/application/dto/ChangePriceRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.application;
+package kitchenpos.products.application.dto;
 
 public class ChangePriceRequest {
     private Long price;
