+++ src/main/java/kitchenpos/products/ui/ProductRestController.java
@@ -1,7 +1,7 @@
 package kitchenpos.products.ui;
 
-import kitchenpos.products.application.ChangePriceRequest;
-import kitchenpos.products.application.CreateProductRequest;
+import kitchenpos.products.application.dto.ChangePriceRequest;
+import kitchenpos.products.application.dto.CreateProductRequest;
 import kitchenpos.products.application.ProductService;
 import kitchenpos.products.domain.Product;
 import org.springframework.http.ResponseEntity;
